# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/submissions": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new submission for a step",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "submissions"
                ],
                "summary": "Create a new submission",
                "parameters": [
                    {
                        "description": "Submission object to create",
                        "name": "submission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SubmissionCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Submission created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Submission"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/submissions/upload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Uploads a file to Google Cloud Storage for a submission",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "submissions"
                ],
                "summary": "Upload a file for a submission",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "team_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "File uploaded successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "url": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.JSONResponse": {
            "type": "object",
            "additionalProperties": true
        },
...
        "models.Evaluation": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.User"
                },
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "note": {
                    "type": "integer"
                },
                "team": {
                    "description": "Belongs to Team",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Team"
                        }
                    ]
                },
                "teamID": {
                    "description": "Foreign key referencing Team.ID",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "description": "Foreign key referencing User.ID, column name as \"author_id\"",
                    "type": "integer"
                }
            }
        },
        "models.File": {
            "type": "object",
            "properties": {
                "contentType": {
                    "description": "File content type",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "hackathon_id": {
                    "description": "Foreign key for Hackathon",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lastModified": {
                    "description": "Last modified date",
                    "type": "string"
                },
                "name": {
                    "description": "Filename",
                    "type": "string"
                },
                "size": {
                    "description": "File size in bytes",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "description": "URL to download the file from Google Cloud Storage",
                    "type": "string"
                },
                "userId": {
                    "description": "User ID",
                    "type": "string"
                }
            }
        },
        "models.Hackathon": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "$ref": "#/definitions/models.User"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "hackathon_files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.File"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "max_participants": {
                    "type": "integer"
                },
                "max_participants_per_team": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nb_of_teams": {
                    "type": "integer"
                },
                "participations": {
                    "description": "Many-to-many relationship with User through Participation",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Participation"
                    }
                },
                "start_date": {
                    "type": "string"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Team"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Participation": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "hackathonID": {
                    "description": "Foreign key referencing Hackathon.ID",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isOrganizer": {
                    "description": "Indicates if the user is an organizer of the hackathon",
                    "type": "boolean"
                },
                "team_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "description": "Foreign key referencing User.ID",
                    "type": "integer"
                }
            }
        },
...
        "models.Skill": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "description": "Many-to-many relationship with User",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.Step": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dead_line_date": {
                    "type": "string",
                    "example": "2025-12-31T23:59:59Z"
                },
                "deletedAt": {
                    "type": "string"
                },
                "hackathon": {
                    "$ref": "#/definitions/models.Hackathon"
                },
                "hackathon_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer"
                },
                "position": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "done"
                },
                "title": {
                    "type": "string",
                    "example": "Step 1"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Submission": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "evaluation": {
                    "$ref": "#/definitions/models.Evaluation"
                },
                "evaluation_id": {
                    "type": "integer"
                },
                "file_url": {
                    "type": "string"
                },
                "git_link": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "step": {
                    "$ref": "#/definitions/models.Step"
                },
                "step_id": {
                    "type": "integer"
                },
                "team": {
                    "$ref": "#/definitions/models.Team"
                },
                "team_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.SubmissionCreate": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "file_url": {
                    "type": "string"
                },
                "git_link": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "team": {
                    "$ref": "#/definitions/models.Team"
                },
                "team_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Team": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "evaluation": {
                    "description": "Relation avec Evaluation",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Evaluation"
                        }
                    ]
                },
                "evaluation_id": {
                    "description": "Ajout de l'EvaluationID",
                    "type": "integer"
                },
                "hackathon": {
                    "$ref": "#/definitions/models.Hackathon"
                },
                "hackathon_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "example": "Team 1"
                },
                "nbOfMembers": {
                    "type": "integer"
                },
                "submission": {
                    "description": "Assuming there is a Submission model",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Submission"
                        }
                    ]
                },
                "submission_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "description": "Has many Users",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password",
                "username"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "$ref": "#/definitions/models.User"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@exmple.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "participations": {
                    "description": "Many-to-many relationship with Hackathon through Participation",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Participation"
                    }
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "profile_picture": {
                    "type": "string"
                },
                "roles": {
                    "description": "0 = user, 2 = organizer, 4 = admin",
                    "type": "integer",
                    "example": 0
                },
                "skills": {
                    "description": "Many-to-many relationship with Skill",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Skill"
                    }
                },
                "team": {
                    "description": "Belongs to Team",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Team"
                        }
                    ]
                },
                "team_id": {
                    "description": "Foreign key referencing Team.ID",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "example": "jdoe"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}